+++ entrypoints/options/App.tsx
@@ -1,0 +1,8 @@
+export default function OptionsPage() {
+  return (
+    <div>
+      option page
+      <ModeToggle />
+    </div>
+  );
+}