+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec2.0.0.html).
 
+## v1.2.6 - Dec 18, 2025
+
+### Added
+
+- Add option page
+- Add anti-FOUC
+- Add dark mode
+- Create popup header component
+
+### Changed
+
+- Update popup design
+- Update app.tsx for popup
+- Update era selector in popup with shadcn ui
+- Update building selector in popup with shadcn ui
+
 ## v1.2.4 - Dec 17, 2025
 
 ### Fixed